Updated evidence system

+++ [scripts]/[jobs]/[government]/evidences/server/biometrics/linked_biometrics.lua
@@ -1,6 +1,9 @@
+local config <const> = require "config"
+local framework <const> = require "common.frameworks.framework"
 local database <const> = require "server.database"
 local citizens <const> = require "server.citizens.citizens"
 local allowedTypes <const> = { "fingerprint", "dna" }
+local logger <const> = require "server.logger"
 
 local linkedBiometrics = {}
 
@@ -25,15 +28,24 @@ MySQL.update.await(
 )
 
 lib.callback.register("evidences:linkBiometricDataToIdentifier", function(source, arguments)
+    if not framework.hasPermission(config.permissions.access, source) then
+        return {
+            success = false,
+            response = "laptop.notifications.no_permission.description"
+        }
+    end
+
     local type <const> = arguments.type
     local biometricData <const> = arguments.biometricData
     local identifier <const> = arguments.identifier
-    
+
     if lib.table.contains(allowedTypes, type) then
         local result <const> = database.update(string.format("DELETE FROM linked_%s WHERE identifier = ?", type), identifier)
 
         if result.success then
             if biometricData then
+                logger.log(source, "Biometric data linked to citizen", { evidenceType = type, biometricData = biometricData, identifier = identifier })
+
                 return database.update(string.format([[
                     INSERT INTO linked_%s (%s, identifier) VALUES (?, ?)
                     ON DUPLICATE KEY UPDATE identifier = ?
@@ -61,10 +73,24 @@ function linkedBiometrics.getCitizenLinkedToBiometricData(type, biometricData)
 end
 
 lib.callback.register("evidences:getCitizenLinkedToBiometricData", function(source, arguments)
+    if not framework.hasPermission(config.permissions.access, source) then
+        return {
+            success = false,
+            response = "laptop.notifications.no_permission.description"
+        }
+    end
+
     return linkedBiometrics.getCitizenLinkedToBiometricData(arguments.type, arguments.biometricData)
 end)
 
 lib.callback.register("evidences:getBiometricDataLinkedToIdentifier", function(source, arguments)
+    if not framework.hasPermission(config.permissions.access, source) then
+        return {
+            success = false,
+            response = "laptop.notifications.no_permission.description"
+        }
+    end
+
     return database.selectFirstRow(
         [[
             SELECT lf.fingerprint, ld.dna
