+++ [scripts]/[jobs]/[government]/evidences/server/citizens/citizens.lua
@@ -1,6 +1,7 @@
 local config <const> = require "config"
 local framework <const> = require "common.frameworks.framework"
 local database <const> = require "server.database"
+local logger <const> = require "server.logger"
 
 require "server.citizens.notes"
 
@@ -10,7 +11,7 @@ if not config.citizens.synced then
     MySQL.update.await(
         [[
             CREATE TABLE IF NOT EXISTS citizens (
-                identifier VARCHAR(500) PRIMARY KEY DEFAULT UUID(),
+                identifier VARCHAR(500) PRIMARY KEY DEFAULT (UUID()),
                 fullName TEXT NOT NULL,
                 birthdate TEXT NOT NULL,
                 gender ENUM('male', 'female', 'non_binary') NOT NULL
@@ -19,6 +20,13 @@ if not config.citizens.synced then
     )
 
     lib.callback.register("evidences:storeCitizen", function(source, arguments)
+        if not framework.hasPermission(config.permissions.access, source) then
+            return {
+                success = false,
+                response = "laptop.notifications.no_permission.description"
+            }
+        end
+
         if arguments.identifier then
             return database.update(
                 [[
@@ -29,57 +37,76 @@ if not config.citizens.synced then
                     WHERE identifier = ?
                 ]],
                 arguments.fullName, arguments.birthdate, arguments.gender, arguments.identifier,
-                function() return arguments end
+                function()
+                    logger.log(source, "Citizen updated", arguments)
+                    return arguments
+                end
             )
         end
 
         return database.selectFirstColumn(
             [[
-                INSERT INTO citizens (fullName, birthdate, gender)
-                VALUES (?, ?, ?)
+                INSERT INTO citizens (identifier, fullName, birthdate, gender)
+                VALUES (UUID(), ?, ?, ?)
                 RETURNING identifier
             ]],
             arguments.fullName, arguments.birthdate, arguments.gender,
             function(identifier)
                 arguments.identifier = identifier
+                logger.log(source, "Citizen created", arguments)
                 return arguments
             end
         )
     end)
 
     lib.callback.register("evidences:deleteCitizen", function(source, arguments)
+        if not framework.hasPermission(config.permissions.access, source) then
+            return {
+                success = false,
+                response = "laptop.notifications.no_permission.description"
+            }
+        end
+
         local identifier <const> = arguments.identifier
 
         local result = database.update("DELETE FROM linked_fingerprint WHERE identifier = ?", identifier)
         if result.success then
             result = database.update("DELETE FROM linked_dna WHERE identifier = ?", identifier)
+
             if result.success then
+                logger.log(source, "Citizen deletion requested", arguments)
                 return database.update("DELETE FROM citizens WHERE identifier = ?", identifier)
             end
         end
-        
+
         return result
     end)
 end
 
 lib.callback.register("evidences:getCitizens", function(source, arguments)
+    if not framework.hasPermission(config.permissions.access, source) then
+        return {
+            success = false,
+            response = "laptop.notifications.no_permission.description"
+        }
+    end
+
     arguments.searchText = arguments.searchText or ""
-    arguments.limit = arguments.limit or 1
     arguments.offset = arguments.offset or 0
 
     if config.citizens.synced then
-        return framework.getCitizens(arguments.searchText, arguments.limit, arguments.offset)
+        return framework.getCitizens(arguments.searchText, arguments.offset)
     end
 
-    local pattern <const> = "%" .. arguments.searchText:gsub("\\", "\\\\"):gsub("%%", "\\%%"):gsub("_", "\\_") .. "%"
+    local pattern <const> = "%" .. arguments.searchText:sub(1, 25):gsub("\\", "\\\\"):gsub("%%", "\\%%"):gsub("_", "\\_") .. "%"
 
     return database.query(
         [[
             SELECT * FROM citizens
             HAVING fullName LIKE ?
-            LIMIT ? OFFSET ?
+            LIMIT 10 OFFSET ?
         ]],
-        pattern, arguments.limit, arguments.offset
+        pattern, arguments.offset
     )
 end)
 
@@ -96,6 +123,13 @@ function citizens.getCitizen(identifier)
 end
 
 lib.callback.register("evidences:getCitizen", function(source, arguments)
+    if not framework.hasPermission(config.permissions.access, source) then
+        return {
+            success = false,
+            response = "laptop.notifications.no_permission.description"
+        }
+    end
+
     return citizens.getCitizen(arguments.identifier)
 end)
